# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: microvm
  version: "1.0.0"
  description: |
    Every sandbox is a real virtual machine with its own kernel. Code inside one
    is assumed hostile.

    There are two ways to run code, and the difference is the point:

    - **Sandboxes** are held. You create one, run commands in it, and it is
      yours until you stop it or its TTL elapses. Creation fails when the node
      is full, so backpressure is yours to handle.
    - **Tasks** are queued. You hand work to the fleet and it runs when a slot
      frees anywhere. Submission never fails for capacity.

    Use a sandbox for several commands sharing state; use a task for throughput.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080/v1
    description: A local daemon. Bound to loopback by default, because this API creates VMs that run arbitrary code.
security:
  - bearerAuth: []
...
paths:
  /sandboxes:
    post:
      operationId: createSandbox
      summary: Create a sandbox
      description: |
        Boots a microVM and waits for it to answer. Returns once the sandbox is
        ready to run commands.

        Fails with `capacity_error` when the node has no room. That is
        deliberate: a sandbox is a reservation, so the caller is told
        immediately rather than left waiting. If you would rather wait, submit a
        task.
      tags: [Sandboxes]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SandboxCreateParams' }
      responses:
        '201':
          description: The sandbox is running and ready.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
        '500': { $ref: '#/components/responses/ApiError' }

    get:
      operationId: listSandboxes
      summary: List sandboxes
      description: |
        This node's sandboxes, newest first -- yours only. A sandbox belongs to the
        tenant that created it, and a list that reported every tenant's would hand
        out the IDs every other route is reached by. An admin key sees the whole
        node, because it is the operator's own.

        A stopped sandbox stays listed for a retention window and is then
        dropped, so this is not a history: a sandbox that ran an hour ago may no
        longer be here at all. Anything you need to keep -- the final metering
        above all -- comes from the reply to the delete, which is the only place
        it is ever stated.
      tags: [Sandboxes]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/EndingBefore'
        - name: state
          in: query
          description: Only sandboxes in this state.
          schema: { $ref: '#/components/schemas/SandboxState' }
        - name: tag
          in: query
          description: |
            Only sandboxes carrying this tag, written `key:value`. Repeat the
            parameter to require several: they AND, so `?tag=env:ci&tag=owner:me`
            returns the sandboxes carrying both. The split is at the first colon,
            so a value may contain colons and a key may not.

            The filter is node-local, because the list is: it searches the
            sandboxes this node holds, not the fleet's. On more than one node you
            get this node's answer, and finding a tagged sandbox anywhere means
            asking each node. A tag nobody set is an empty page, not an error --
            there is no registry of known keys to be wrong about.
          explode: true
          schema:
            type: array
            items: { type: string, pattern: '^[^:]+:.*$' }
          example: ["env:ci", "owner:billing-team"]
      responses:
        '200':
          description: A page of sandboxes, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SandboxList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    get:
      operationId: retrieveSandbox
      summary: Retrieve a sandbox
      description: |
        Returns the sandbox with live metering.

        A stopped sandbox is still here for a retention window, so a caller who
        polls learns *why* it stopped -- expired, idle, failed -- rather than
        being handed a bare 404 to interpret. Once the window elapses the record
        is dropped and this returns 404, which is why a 404 on an ID you know
        existed means "too late", not "never was".
      tags: [Sandboxes]
      responses:
        '200':
          description: The sandbox.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }

    delete:
      operationId: deleteSandbox
      summary: Delete a sandbox
      description: |
        Kills the VM and releases everything it held.

        Returns the sandbox with its final metering. Those numbers are sampled
        before the kill and cannot be obtained afterwards: the cgroup that holds
        them dies with the VM, so this reply is the only record of what the
        sandbox cost.
      tags: [Sandboxes]
      responses:
        '200':
          description: The sandbox, stopped, with its final cost.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`. Tokens are compared in constant time.
  parameters:
    SandboxId:
      name: sandbox
      in: path
      required: true
      description: |
        A sandbox belongs to the tenant whose token created it, and every route
        below resolves it against the caller: another tenant's sandbox is a `404`,
        not a `403`. An ID is not a capability, and a `403` would confirm which of a
        guessed range exist. So a `404` here means one of three things -- never
        existed, already forgotten, or not yours -- and the API will not say which.
      schema: { type: string, pattern: '^sb_[0-9A-HJKMNP-TV-Z]{26}$', example: "sb_01JZ8QK3M4N5P6R7S8T9V0W1X2" }
...
    Limit:
      name: limit
      in: query
      description: How many objects to return, 1–100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    StartingAfter:
      name: starting_after
      in: query
      description: |
        A cursor. Returns the page of objects immediately after this ID, which
        is the ID of the last object on the previous page.
      schema: { type: string }
    EndingBefore:
      name: ending_before
      in: query
      description: |
        A cursor, for paging backwards. Returns the page immediately before this
        ID, which is the ID of the first object on the current page.
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Makes a retry safe. A request that never got its reply -- a timeout, a
        dropped connection -- cannot be known to have failed, so the only sane
        move is to retry it; without a key, that retry runs the work twice.

        The first request with a given key is executed and its reply recorded.
        Any later request with the same key replays that reply, without doing
        the work again. Keys are held for 24 hours and are scoped to the token.

        Reusing a key with a *different* request body is an `idempotency_error`:
        it means the client believes it is retrying something it is not.
      schema: { type: string, maxLength: 255, example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890" }
  responses:
    InvalidRequest:
      description: The request is malformed or contradicts itself.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: The token is missing or wrong.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: No such object.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: The object already exists, or is in a state that forbids this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    CapacityExhausted:
      description: |
        No room right now. Either the node has no free slot, or the caller is over
        a per-tenant limit -- both are `capacity_error` with code
        `node_at_capacity`, deliberately, because the caller's move is the same and
        that is the pair every SDK already backs off on. The message is what tells
        them apart: a full node frees up on its own, whereas being over your own
        sandbox cap needs you to delete one.
      headers:
        Retry-After:
          description: |
            Seconds to wait, on the refusals whose end can be named to the second --
            the request-rate limit. Absent for a full node and for the per-tenant
            sandbox cap: when a slot frees is not something this node knows, and a
            number invented here would be a guess dressed as a fact.
          required: false
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ApiError:
      description: Something failed on our side.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      description: The body of every non-2xx reply.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }
...
    SandboxState:
      type: string
      enum: [running, stopped]
    SandboxStopReason:
      type: string
      description: |
        Why a sandbox stopped. `stopped` means you asked; the rest happened to
        it, and only `failed` means something went wrong.
      enum:
        - stopped   # deleted by the caller
        - expired   # its TTL elapsed
        - idle      # nothing ran for its idle timeout
        - failed    # the VM died
    Sandbox:
      type: object
      required: [id, object, image, state, created, expires, stats]
      properties:
        id: { type: string, example: "sb_01JZ8QK3M4N5P6R7S8T9V0W1X2" }
        object: { type: string, enum: [sandbox] }
        image: { type: string, example: "python" }
        state: { $ref: '#/components/schemas/SandboxState' }
        stop_reason:
          allOf: [{ $ref: '#/components/schemas/SandboxStopReason' }]
          description: Set once the sandbox is stopped.
        created: { type: string, format: date-time }
        expires:
          type: string
          format: date-time
          description: |
            When the TTL kills it, whatever it is doing. Movable with
            `POST /sandboxes/{sandbox}/extend`, but only up to the host's maximum
            lifetime counted from `created`, so this field can be pushed back and
            never removed.
        stopped: { type: string, nullable: true, format: date-time }
        vcpus: { type: integer }
        mem_mib: { type: integer }
        network:
          type: boolean
          description: Whether filtered egress is on.
        tags:
          type: object
          additionalProperties: { type: string }
          description: |
            The labels set at create time, returned exactly as they were given.

            Unlike `env`, tags **are** returned -- here, and in every list that
            includes this sandbox. Never put a credential in one. A secret that
            needs to reach the guest is what `env` is for, and `env` is
            write-only precisely so this mistake cannot be made by accident.
          example: {"env": "ci", "owner": "billing-team"}
        storage:
          allOf: [{ $ref: '#/components/schemas/SandboxStorage' }]
          description: |
            Where this sandbox's files persist, when the node has object storage
            configured. Absent when it does not.
        stats: { $ref: '#/components/schemas/Stats' }
...
    SandboxList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Sandbox' }
...
    SandboxCreateParams:
      type: object
      required: [image]
      properties:
        image:
          type: string
          description: The language image.
          example: "python"
        vcpus: { type: integer, minimum: 1, default: 1 }
        mem_mib: { type: integer, minimum: 128, default: 256 }
        cpu_cores:
          type: number
          format: double
          description: A hard CPU ceiling; may be fractional. 0.5 means half a core, enforced.
          example: 0.5
        disk_mib: { type: integer }
        network_bps:
          type: integer
          description: |
            Caps bandwidth both ways, in bytes per second. Worth setting even
            when CPU is capped: a sandbox pinned to a quarter core can still
            saturate the host's uplink, and the first you hear of it is the
            abuse complaint.
        disk_bps: { type: integer }
        disk_iops: { type: integer }
        network:
          type: boolean
          default: false
          description: |
            Filtered egress. The public internet is reachable; every private
            range, link-local and the cloud metadata endpoint are not.
        env:
          type: object
          additionalProperties: { type: string }
          description: |
            Injected into every execution in this sandbox. An execution's own
            env wins where they overlap.

            Write-only. These usually carry credentials, so they are never
            logged and no endpoint returns them. They are applied by the host on
            each execution rather than written into the guest's filesystem,
            where anything inside could read them long after the command that
            needed them.
        tags:
          type: object
          additionalProperties: { type: string }
          maxProperties: 10
          description: |
            Labels of your own, for finding this sandbox again: filter
            `GET /sandboxes` by `tag`, or read them back off the object.

            At most 10 pairs, keys at most 64 bytes and values at most 256. Bytes
            rather than characters, so a key written in emoji runs out about four
            times sooner than its length suggests. A key may not be empty and may
            not contain `:`, which is the character the `tag` filter splits on.

            Node-local, like the list they are filtered on: they live in the
            daemon's memory beside the sandbox and go when it is forgotten. They
            are not written into the guest and nothing inside it can read them,
            which is what separates them from `env` -- a tag names a sandbox, it
            does not configure one.

            **Returned by the API, unlike `env`.** Anything that can read this
            sandbox can read its tags -- which is its own tenant, and an admin key
            -- so a tag is a label and never a secret.
          example: {"env": "ci", "owner": "billing-team"}
        ttl_seconds:
          type: integer
          default: 900
          description: |
            Maximum lifetime. The sandbox is killed when it elapses. Capped by
            the host's own maximum, and extensible later within that cap with
            `POST /sandboxes/{sandbox}/extend`.
        idle_seconds:
          type: integer
          default: 120
          description: Reclaim after this long with nothing running. Negative disables it.
        storage:
          allOf: [{ $ref: '#/components/schemas/SandboxStorageParams' }]
          description: |
            Options for this sandbox's storage mount. The namespace is never set
            here — it comes from your API key. Only the read/write posture is
            yours to choose, and only to tighten.
    SandboxStorageParams:
      type: object
      description: What a create request may say about storage, which is very little.
      properties:
        read_only:
          type: boolean
          description: |
            Mount storage read-only for this sandbox even if the key could write.
            It can only tighten: a read-only key stays read-only whatever you set
            here. Useful for a run that should read shared state but never touch it.
        mount_path:
          type: string
          description: |
            Where to mount storage inside the guest. An absolute, clean path using
            only letters, digits and `_-./`; it names a directory in your own
            sandbox and cannot reach another tenant's files. Defaults to
            `/mnt/storage`.
          example: "/data"
...